Add connection tune ok frame decoding

+++ broker-amqp/src/main/java/org/wso2/broker/amqp/codec/data/EncodableData.java
@@ -25,7 +25,7 @@ import io.netty.buffer.ByteBuf;
  *
  */
 public interface EncodableData {
-    int getSize();
+    long getSize();
 
     void write(ByteBuf buf);
 }
